refactor: move DoppelgangerState to private implementation
- Remove state.rs module and keep DoppelgangerState private in service.rs
- Simplify tests to use public API (is_monitoring) instead of internal state comparisons
- Remove redundant assert_eq checks in favor of behavior-focused assertions
- Fix clippy lint for identity operation in test
