+++ anchor/operator_doppelganger/src/lib.rs
@@ -1,4 +1,3 @@
 mod service;
-mod state;
 
 pub use service::OperatorDoppelgangerService;
