+++ anchor/operator_doppelganger/src/service.rs
@@ -7,7 +7,74 @@ use ssv_types::{consensus::QbftMessage, message::SignedSSVMessage};
 use task_executor::{ShutdownReason, TaskExecutor};
 use tracing::{error, info};
 
-use super::state::DoppelgangerState;
+/// State of operator doppelgänger detection
+///
+/// ## State Transitions
+///
+/// ```text
+/// GracePeriod → Monitoring → Completed
+/// ```
+///
+/// - **GracePeriod**: Waiting for network message caches to expire before checking
+/// - **Monitoring**: Actively checking messages for doppelgängers
+/// - **Completed**: Monitoring period finished, no longer checking
+#[derive(Debug, Clone, PartialEq, Eq)]
+enum DoppelgangerState {
+    /// In startup grace period - not yet checking for doppelgängers
+    ///
+    /// ## Why we need a grace period
+    ///
+    /// Gossipsub stores messages in a sliding window cache (mcache) for gossip propagation.
+    /// Messages remain in this cache for `history_length × heartbeat_interval` (~4.2s in Anchor).
+    ///
+    /// **The restart vulnerability:**
+    /// 1. Node sends messages (QBFT + partial signatures) at t=0
+    /// 2. Messages propagate to peers' mcache
+    /// 3. Node crashes/restarts at t=2s
+    /// 4. Gossipsub seen_cache is cleared (not persisted)
+    /// 5. Messages still in peers' mcache (~2s remaining)
+    /// 6. Node reconnects and receives its own messages via IHAVE/IWANT
+    /// 7. FALSE POSITIVE: Messages have our operator ID but we think they're from a twin!
+    ///
+    /// **Solution:**
+    /// Wait for gossip cache expiry (~5s) before checking messages. This ensures our own
+    /// old messages have expired from the network before we start detecting twins.
+    GracePeriod,
+
+    /// Actively monitoring for doppelgängers - checking all messages
+    Monitoring,
+
+    /// Monitoring period completed - no longer checking for doppelgängers
+    Completed,
+}
+
+impl Default for DoppelgangerState {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl DoppelgangerState {
+    /// Create a new doppelgänger state starting in grace period
+    fn new() -> Self {
+        Self::GracePeriod
+    }
+
+    /// Check if actively monitoring
+    fn is_monitoring(&self) -> bool {
+        matches!(self, Self::Monitoring)
+    }
+
+    /// Transition from grace period to monitoring
+    fn end_grace_period(&mut self) {
+        *self = Self::Monitoring;
+    }
+
+    /// Transition from monitoring to completed
+    fn end_monitoring(&mut self) {
+        *self = Self::Completed;
+    }
+}
 
 pub struct OperatorDoppelgangerService {
     /// Our operator ID to watch for (wraps database watch)
@@ -301,12 +368,27 @@ mod tests {
     }
 
     #[test]
+    fn test_state_lifecycle() {
+        let mut state = DoppelgangerState::new();
+
+        // Start in grace period - not monitoring
+        assert!(!state.is_monitoring());
+
+        // Transition to monitoring
+        state.end_grace_period();
+        assert!(state.is_monitoring());
+
+        // Complete monitoring
+        state.end_monitoring();
+        assert!(!state.is_monitoring());
+    }
+
+    #[test]
     fn test_service_creation() {
         let service = create_service();
 
         // Start in grace period, not yet monitoring
         assert!(!service.is_monitoring());
-        assert!(service.state.lock().is_grace_period());
     }
 
     // High-value tests for check_message functionality
@@ -321,7 +403,7 @@ mod tests {
         spawn_and_advance_past_grace_period(service.clone(), &executor).await;
 
         // Grace period should be complete, now monitoring
-        assert!(service.state.lock().is_monitoring());
+        assert!(service.is_monitoring());
 
         // Create a single-signer message with our operator ID (1)
         let (signed_message, qbft_message) =
@@ -349,7 +431,7 @@ mod tests {
             .spawn_monitor_task(grace_period, wait_epochs, &executor);
 
         // Still in grace period (don't advance time)
-        assert!(service.state.lock().is_grace_period());
+        assert!(!service.is_monitoring());
 
         // Create a single-signer message with our operator ID (1)
         let (signed_message, qbft_message) =
@@ -426,7 +508,7 @@ mod tests {
         tokio::task::yield_now().await;
 
         // Calculate monitoring duration
-        let monitoring_duration = Duration::from_secs(wait_epochs * 1 * 12); // epochs * slots_per_epoch * slot_duration
+        let monitoring_duration = Duration::from_secs(wait_epochs * 12); // epochs * (slots_per_epoch=1) * (slot_duration=12s)
 
         // Advance time past grace period first
         tokio::time::advance(grace_period).await;
